refactor: GameResult를 삭제하고 책임을 DealerResult로 이전, State를 단독으로 분리
