+++ src/main/kotlin/blackjack/model/DealerResult.kt
@@ -1,3 +1,3 @@
 package blackjack.model
 
-data class DealerResult(val map: Map<GameResult.State, Int>)
+data class DealerResult(val map: Map<WinningState, Int>)
