+++ src/main/kotlin/blackjack/model/PlayerResult.kt
@@ -1,3 +1,3 @@
 package blackjack.model
 
-data class PlayerResult(val name: String, val state: GameResult.State)
+data class PlayerResult(val name: String, val winningState: WinningState)
